traffic_anmly_analysis
affic signals are part of our critical infrastructure and protecting their integrity is a serious concern. Security flaws in traffic signal systems have been documented and effective detection
of exploitation of these flaws remains a challenge. In this paper
we present a visual analytics approach to look for anomalies in
traffic signal data (i.e., abnormal traffic light patterns) that may
indicate a compromise of the system. To our knowledge it is a
first time a visual analytics approach is applied for the processing and exploration of traffic signal data. This system supports
level-of-detail exploration with various visualization techniques.
Data cleaning and a number of preprocessing techniques for the
extraction of summary information (e.g., traffic signal cycles) of
the data are also performed before the visualization and data exploration. Our system successfully reveals the errors in the input
data that would be difficult to capture with simple plots alone. In
addition, our system captures some abnormal signal patterns that
may indicate intrusions into the system. In summary, this work
offers a new and effective way to study attacks or intrusions to
traffic signal control systems via the visual analysis of traffic light
signal patterns.